⏪️ Revert "Merge pull request #2841 from nextcloud/feature/heading-anchors" until #2868 be fixed
This reverts commit 0f2c48ea4c4822456f737b3c39b0c5f354b17df5, reversing
changes made to 9973d1fb5230d8c9fb46394a14fbdcc90d339a3a.
